requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
flatbuffers==24.3.25
idna==3.6
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.3
packaging==24.1
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.2.0
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
wcwidth==0.2.13
